// backend/src/common/enums.ts
export enum BillingCycle {
  Monthly = "monthly",
  Yearly = "yearly",
}

export enum PaymentType {
  Visa = "visa",
  Mastercard = "mastercard",
  Amex = "amex",
  Bank = "bank",
  PromptPay = "promptpay",
  TrueMoney = "truemoney",
}

export enum Category {
  Entertainment = "entertainment",
  Productivity = "productivity",
  Developer = "developer",
  Cloud = "cloud",
}